move imports & ImportError check inside ImageToTensorBoard class
